Module wasm_rdfshape_validator::app::api

Proporciona funciones y estructuras necesarias para realizar solicitudes HTTP, manejar las respuestas y procesar los datos recibidos.

Structs

ApiResult

Define el esquema para el objeto Result de la respuesta del método Validation

Data

Objeto Data de la respuesta al andpoint validation

InfoRdfRequest

Define el esquema para el body del método data/info

InfoRdfRequestContent

Define el esquema para el objeto data del body del método data/info

InfoRdfResponse

Define el esquema para la respuesta del método data/info

InfoRdfResult

Define el esquema objeto result para la respuesta del método data/info

InfoShexRequest

Define el esquema para el body del método schema/info

InfoShexRequestContent

Define el esquema para el objeto schema el body del método schema/info

InfoShexResponse

Define el esquema la respuesta del método schema/info

InfoShexResult

Define el esquema el objeto result la respuesta del método schema/info

Prefix

Define el esquema el objeto Prefix la respuesta del método schema/info

RequestBody

Define el esquema para el body de Validation

Schema

Define esquema ShEx

ShapeMap

Define esquema ShapeMap

ShapeMapEntry

Define el esquema para cada ShapeMap de la respuesta del método Validation

TriggerMode

Define objeto TriggerMode

ValidationResult

Define el esquema para la respuesta del método Validation

Functions

call_rdf_info_api

Realiza una solicitud a la API para obtener información sobre RDF.

call_shex_info_api

Realiza una solicitud a la API para obtener información sobre ShEx.

call_validation_api

Realiza una llamada a la API para validar RDF, ShEx y ShapeMap y maneja los resultados.

create_rdf_info_request_body

Construye el cuerpo de la solicitud para obtener información RDF.

create_shex_info_request_body

Construye el cuerpo de la solicitud para obtener información ShEx.

create_validation_request_body

Construye el cuerpo de la solicitud para validar RDF, ShEx y ShapeMap.

extract_last_segment 🔒

Extrae la última parte de una URI, usada para simplificar las referencias en las visualizaciones.

format_shape_maps

Formatea las entradas de ShapeMap para mejorar la legibilidad en la visualización final.

format_status 🔒

Formatea el estado de la validación para ser mostrado de manera clara.