Enum wasm_rdfshape_validator::app::Msg

pub enum Msg {
    Validate,
    ValidationResult(ValidationResult, String),
    UpdateShapeMapValue(String),
    LoadExample(String),
    ExampleLoaded(Result<ExampleData, String>),
    CloseAlert,
    OpenModal(String, String),
    CloseModal,
}

Enum para los mensajes que se pueden enviar al componente

Variants

Validate

Solicita la validación de los datos actualmente cargados en los editores.

ValidationResult(ValidationResult, String)

Resultado de una solicitud de validación. Contiene api::ValidationResult con el resultado de la validación y un String que puede contener un mensaje de error.

UpdateShapeMapValue(String)

Actualiza el valor actual del ShapeMap con el nuevo valor proporcionado. String contiene el nuevo valor de ShapeMap.

LoadExample(String)

Carga un ejemplo específico. String contiene el identificador del ejemplo a cargar.

ExampleLoaded(Result<ExampleData, String>)

Notifica que un ejemplo ha sido cargado. Result<ExampleData, String> contiene los datos del ejemplo o un mensaje de error en caso de fallo.

CloseAlert

Cierra la alerta actualmente mostrada, por ejemplo, un mensaje de error.

OpenModal(String, String)

Abre un modal con información detallada. String, String contiene el título y el contenido del modal, respectivamente.

CloseModal

Cierra el modal actualmente abierto.